requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
dj-database-url
drf-spectacular